+++ operator.ps1
@@ -205,6 +205,7 @@ if ( $IsWindows ){
 . $pwshSrcDir/Invoke-Link_function.ps1
 
 ## misc
+. $pwshSrcDir/Test-isAsciiLine_function.ps1
 . $pwshSrcDir/Set-NowTime2Clipboard_function.ps1
 . $pwshSrcDir/tenki_function.ps1
 . $pwshSrcDir/sleepy_function.ps1